This renames the namespace from the original com.github.barteksc -> com.github.polesapart
This is not meant to claim authorship or whatever, I simply thought it would
be required to make publishin in jitpack.io works - but it seems
I was mistaken.

If there are any reasons I can consider rolling back in the future.  Anyway,
as the #743 broke the ABI slightly, there's no point in being too strict
about it right now.